Fixed the values of the Radius
Items would spawn way beyond the radius of the arena. Fixed that, and now i'm on to the Machine Learning portion. Can't get past the actual implementation of the ML due to something not being properly assigned, or the Unity software is taking too long. Not really sure what i'm doing wrong